Remove DLQ route integration test, add happy path to unit tests
